Fixed signedness casts.

+++ libfreerdp/core/gcc.c
@@ -1671,7 +1671,7 @@ BOOL gcc_read_client_cluster_data(wStream* s, rdpMcs* mcs, UINT16 blockLength)
 
 	if (blockLength != 8)
 	{
-		if (Stream_GetRemainingLength(s) >= (blockLength - 8))
+		if (Stream_GetRemainingLength(s) >= (size_t)(blockLength - 8))
 		{
 			/* The old Microsoft Mac RDP client can send a pad here */
 			Stream_Seek(s, (blockLength - 8));
@@ -1740,7 +1740,7 @@ BOOL gcc_read_client_monitor_data(wStream* s, rdpMcs* mcs, UINT16 blockLength)
 		monitorCount = settings->MonitorDefArraySize;
 	}
 
-	if (((blockLength - 8)  / 20) < monitorCount)
+	if ((UINT32)((blockLength - 8)  / 20) < monitorCount)
 		return FALSE;
 
 	settings->MonitorCount = monitorCount;
